requirements.txt
streamlit
langchain
langchain-openai
faiss-cpu
openai
sentence-transformers
numpy
python-dotenv
torch
watchdog
